+++ projects/core/migrations/generate-migrations.ts
@@ -14,6 +14,7 @@ import {
 } from './migration-types.js'
 import { isOfType } from '../src/isOfType.js'
 import { initDataProvider } from '../server/initDataProvider.js'
+import { initDataProviderOrJson } from '../server/initDataProviderOrJson'
 
 /**
  * Generates migration scripts based on changes in entities.
@@ -46,7 +47,7 @@ export async function generateMigrations(options: {
     options.snapshotFile || path.join(migrationDir, 'migrations-snapshot.json')
   const migrationsTSFilename =
     options.migrationsTSFile || path.join(migrationDir, 'migrations.ts')
-  const dataProvider = await initDataProvider(options.dataProvider)
+  const dataProvider = await initDataProviderOrJson(options.dataProvider)
 
   for (const p of [snapshotFileName, migrationsTSFilename]) {
     if (!fs.existsSync(path.dirname(p))) {